+++ src/nodes/_specs/emoji.spec.ts
@@ -1,26 +1,24 @@
-import { expect } from 'chai';
+import { expect, use } from 'chai';
 import { Document } from '../../index';
-import { validator, ValidationError, Validator } from '../../_specs/validate';
+import { adfValidator } from '../../_chai';
 import { emoji, Emoji } from '../emoji';
 
 describe('Emoji', () => {
 
-  let validate: Validator;
-
-  function validateEmoji(emo: Emoji) {
+  function docFromEmoji(emo: Emoji) {
     const doc = new Document();
     doc.paragraph().add(emo);
-    validate(doc);
+    return doc;
   }
 
   before(async function() {
     this.timeout(5000);
-    validate = await validator();
+    use(await adfValidator());
   });
 
   it('should create a valid emoji with shortName', () => {
     const emo = emoji('smile');
-    expect(() => validateEmoji(emo)).to.not.throw(ValidationError);
+    expect(docFromEmoji(emo)).to.be.validADF();
     expect(emo.toJSON()).to.deep.equal({
       type: 'emoji',
       attrs: {
@@ -31,7 +29,7 @@ describe('Emoji', () => {
 
   it('should create a valid emoji with shortName and id', () => {
     const emo = emoji('smile', '123');
-    expect(() => validateEmoji(emo)).to.not.throw(ValidationError);
+    expect(docFromEmoji(emo)).to.be.validADF();
     expect(emo.toJSON()).to.deep.equal({
       type: 'emoji',
       attrs: {
@@ -43,7 +41,7 @@ describe('Emoji', () => {
 
   it('should create a valid emoji with shortName, id and text', () => {
     const emo = emoji('smile', '123', ':)');
-    expect(() => validateEmoji(emo)).to.not.throw(ValidationError);
+    expect(docFromEmoji(emo)).to.be.validADF();
     expect(emo.toJSON()).to.deep.equal({
       type: 'emoji',
       attrs: {