+++ src/nodes/_specs/media-group.spec.ts
@@ -1,14 +1,12 @@
-import { expect } from 'chai';
+import { expect, use } from 'chai';
 import { Document } from '../../index';
-import { validator, ValidationError, Validator } from '../../_specs/validate';
+import { adfValidator } from '../../_chai';
 
 describe('Media Group', () => {
 
-  let validate: Validator;
-
   before(async function() {
     this.timeout(5000);
-    validate = await validator();
+    use(await adfValidator());
   });
 
   it('should not allow empty media groups', () => {
@@ -20,7 +18,7 @@ describe('Media Group', () => {
   it('should create valid media group for a file', () => {
     const doc = new Document();
     const group = doc.mediaGroup().file('id', 'coll');
-    expect(() => validate(doc)).to.not.throw(ValidationError);
+    expect(doc).to.be.validADF();
     expect(group.toJSON()).to.deep.equal({
       type: 'mediaGroup',
       content: [
@@ -39,7 +37,7 @@ describe('Media Group', () => {
   it('should create valid media group for a link', () => {
     const doc = new Document();
     const group = doc.mediaGroup().link('id', 'coll');
-    expect(() => validate(doc)).to.not.throw(ValidationError);
+    expect(doc).to.be.validADF();
     expect(group.toJSON()).to.deep.equal({
       type: 'mediaGroup',
       content: [
@@ -63,7 +61,7 @@ describe('Media Group', () => {
       type: 'link',
       occurrenceKey: 'key'
     });
-    expect(() => validate(doc)).to.not.throw(ValidationError);
+    expect(doc).to.be.validADF();
     expect(group.toJSON()).to.deep.equal({
       type: 'mediaGroup',
       content: [