+++ src/nodes/_specs/ordered-list.spec.ts
@@ -1,14 +1,12 @@
-import { expect } from 'chai';
+import { expect, use } from 'chai';
 import { Document } from '../../index';
-import { validator, ValidationError, Validator } from '../../_specs/validate';
+import { adfValidator } from '../../_chai';
 
 describe('Ordered List', () => {
 
-  let validate: Validator;
-
   before(async function() {
     this.timeout(5000);
-    validate = await validator();
+    use(await adfValidator());
   });
 
   it('should not allow empty ordered lists', () => {
@@ -29,7 +27,7 @@ describe('Ordered List', () => {
       const doc = new Document();
       const list = doc.orderedList();
       list.item().paragraph().text('item');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [
@@ -56,7 +54,7 @@ describe('Ordered List', () => {
       const list = doc.orderedList();
       list.item().paragraph().text('item1');
       list.item().paragraph().text('item2');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [
@@ -95,14 +93,20 @@ describe('Ordered List', () => {
     it('should create a nested bullet list', () => {
       const doc = new Document();
       const list = doc.orderedList();
-      list.item().bulletList().item().paragraph().text('nested');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      const item = list.item();
+      item.paragraph().text('line 1');
+      item.bulletList().item().paragraph().text('nested');
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [
           {
             type: 'listItem',
             content: [
+              {
+                type: 'paragraph',
+                content: [{ type: 'text', text: 'line 1' }]
+              },
               {
                 type: 'bulletList',
                 content: [
@@ -122,14 +126,20 @@ describe('Ordered List', () => {
     it('should create a nested ordered list', () => {
       const doc = new Document();
       const list = doc.orderedList();
-      list.item().orderedList().item().paragraph().text('nested');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      const item = list.item();
+      item.paragraph().text('line 1');
+      item.orderedList().item().paragraph().text('nested');
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [
           {
             type: 'listItem',
             content: [
+              {
+                type: 'paragraph',
+                content: [{ type: 'text', text: 'line 1' }]
+              },
               {
                 type: 'orderedList',
                 content: [
@@ -154,7 +164,7 @@ describe('Ordered List', () => {
       list
         .textItem('item1')
         .textItem('item2');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [
@@ -198,7 +208,7 @@ describe('Ordered List', () => {
       list
         .linkItem('Link 1', 'https://example.com/1')
         .linkItem('Link 2', 'https://example.com/2');
-      expect(() => validate(doc)).to.not.throw(ValidationError);
+      expect(doc).to.be.validADF();
       expect(list.toJSON()).to.deep.equal({
         type: 'orderedList',
         content: [