+++ src/nodes/_specs/rule.spec.ts
@@ -1,20 +1,18 @@
-import { expect } from 'chai';
+import { expect, use } from 'chai';
 import { Document } from '../../index';
-import { validator, ValidationError, Validator } from '../../_specs/validate';
+import { adfValidator } from '../../_chai';
 
 describe('Rule', () => {
 
-  let validate: Validator;
-
   before(async function() {
     this.timeout(5000);
-    validate = await validator();
+    use(await adfValidator());
   });
 
   it('should create valid rule', () => {
     const doc = new Document();
     doc.rule();
-    expect(() => validate(doc)).to.not.throw(ValidationError);
+    expect(doc).to.be.validADF();
     expect(doc.toJSON()).to.deep.equal({
       type: 'doc',
       version: 1,